requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.0
gradio>=4.0.0
gradio-pdf>=0.0.1
pymupdf>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.0.0
anthropic>=0.8.0
clip @ git+https://github.com/openai/CLIP.git
timm>=0.9.0
einops>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
accelerate>=0.25.0
safetensors>=0.4.0
